+++ include/bsp/iotc_bsp_crypto.h
@@ -1,6 +1,6 @@
-/* Copyright 2018 Google LLC
+/* Copyright 2018-2019 Google LLC
  *
- * This is part of the Google Cloud IoT Edge Embedded C Client,
+ * This is part of the Google Cloud IoT Device SDK for Embedded C,
  * it is licensed under the BSD 3-Clause license; you may not use this file
  * except in compliance with the License.
  *
@@ -22,15 +22,20 @@
 
 #include "iotc_types.h"
 
+#ifdef __cplusplus
+extern "C" {
+#endif
+
 /**
  * @file iotc_bsp_crypto.h
- * @brief IoTC Client's Board Support Package (BSP) for using a Cryptography
- * library implementation to encode and sign portions of clients Google Cloud
- * IoT Core JWT credential.
+ * @brief The IoT Device SDK Board Support Package (BSP) for using a
+ * Cryptography library implementation to encode and sign portions of clients
+ * Google Cloud IoT Core JWT credential.
  *
  * This file defines the API of a Crypto Board Support Package (BSP) that the
  * IoTC Client would use to generate and sign a JWT using ECC keys and SHA256
- * digests, and formatting results in a URL-Safe Base64 string.
+ * digests, and formatting results in a URL-Safe Base64 string. Reference
+ * implementations reside in the src/bsp/crypto directory of this repository.
  *
  * The BSP has been created in a way that allows for flexibility in
  * implementation. The private key data and/or private key slot number
@@ -39,7 +44,8 @@
  * implementation of this BSP.
  *
  * Note that this implementation need not worry about how to format the JWT
- * itself, as this has been already done by the IoTC Client internal library.
+ * itself, as this has been already done by the IoT Device SDK in a previous
+ * step.
  *
  * These functions should return one of the Crypto BSP enumerated states
  * defined in iotc_bsp_crypto_state_t to report success or failures.
@@ -155,8 +161,12 @@ iotc_bsp_crypto_state_t iotc_bsp_sha256(uint8_t* dst_buf_32_bytes,
  * iotc_bsp_crypto_state_t values, when applicable.
  */
 iotc_bsp_crypto_state_t iotc_bsp_ecc(
-    const iotc_crypto_private_key_data_t* private_key_pem, uint8_t* dst_buf,
+    const iotc_crypto_key_data_t* private_key_pem, uint8_t* dst_buf,
     size_t dst_buf_size, size_t* bytes_written, const uint8_t* src_buf,
     size_t src_buf_size);
+
+#ifdef __cplusplus
+}
+#endif
 
 #endif /* __IOTC_BSP_CRYPTO_H__ */